alpha-pod/controller/requirements.txt
kafka-python==2.0.2
prometheus-api-client==0.4.2
kubernetes==17.17.0